requirements.txt
Flask==3.0.0
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.1.0
scikit-learn==1.3.2
matplotlib==3.8.2
Werkzeug==3.0.1


